backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
blinker==1.9.0
click==8.3.0
fastapi==0.119.1
Flask==3.1.2
flask-ldap3-login==1.0.2
Flask-Login==0.6.3
Flask-WTF==1.2.2
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
ldap3==2.9.1
libvirt-python==11.8.0
MarkupSafe==3.0.3
pyasn1==0.6.1
pydantic==2.12.3
pydantic_core==2.41.4
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
Werkzeug==3.1.3
WTForms==3.2.1
python-dotenv==1.0.1
